Replace xrange with range. (#17)
In Python 3, range is the same as Python 2's xrange.  For small ranges, there isn't much benefit to using a generator instead of just generating a list, so it is simpler to just use range in both versions of Python instead of switching between how it is called depending on Python version when the ranges used are small. In this case, most are smaller than 201 entries, most appear in only test/example code, and it makes no noticable difference in how long tests take to run.